pyproject.toml
[tool.poetry]
name = "spotify-podcast"
version = "0.1.0"
description = "Spotify Podcast API"
authors = ["Sohret Aga <shohretaga@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2.5"
djangorestframework = "^3.16.1"
python-dotenv = "^1.1.1"
requests = "^2.32.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
